+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v1.5.2 (2025-06-13)
+
+### Fixed bugs
+* 
+
+### Enhancements
+* Added security context fields to ContainerImage allowing for `runAsUser`, `runAsGroup`, `allowPrivilegeEscalation` and `readOnlyRootFilesystem` to be set at container level
+
+### Upgrade guide
+1. Deploy new image
+2. Migrate database [(Guide)](https://github.com/4spacesdk/kubernetes-service-orchestrator?tab=readme-ov-file#migrate-database-helm)
+ 
+
+
 ## v1.5.1 (2025-06-07)
 
 ### Fixed bugs
@@ -7,7 +21,7 @@
 * Add imagePullSecrets to CronJobs [Issue #51](https://github.com/4spacesdk/kubernetes-service-orchestrator/issues/51)
 
 ### Enhancements
-* 
+*
 
 ### Upgrade guide
 1. Deploy new image